fix(guard): tighten passive trust backend coverage
