+++ tests/test_guard_policy_integrity.py
@@ -1718,119 +1718,6 @@ def test_trust_cli_explain_reports_guard_cloud_rule(tmp_path: Path, capsys) -> N
     assert "integrity_status" not in payload["rule"]
 
 
-def test_trust_cli_reports_unsupported_backend_status_without_prompt_error(tmp_path: Path, capsys) -> None:
-    home_dir = tmp_path / "home"
-    rc = main(
-        [
-            "guard",
-            "trust",
-            "status",
-            "--backend",
-            "macos-native",
-            "--home",
-            str(home_dir),
-            "--json",
-        ]
-    )
-    payload = json.loads(capsys.readouterr().out)
-
-    assert rc == 0
-    assert payload["mode"] == "unsupported"
-    assert payload["backend_requested"] == "macos-native"
-    assert payload["message"].startswith("This local trust backend is unsupported")
-    assert payload["passive_prompt_allowed"] is False
-
-
-def test_trust_cli_degraded_safe_backend_is_explicit(tmp_path: Path, capsys) -> None:
-    home_dir = tmp_path / "home"
-    rc = main(
-        [
-            "guard",
-            "trust",
-            "status",
-            "--backend",
-            "degraded-safe",
-            "--home",
-            str(home_dir),
-            "--json",
-        ]
-    )
-    payload = json.loads(capsys.readouterr().out)
-
-    assert rc == 0
-    assert payload["backend_requested"] == "degraded-safe"
-    assert payload["backend"] == "degraded-safe"
-    assert payload["remembered_rules"] == "disabled_degraded"
-    assert payload["durable_local_rules"] == "limited"
-
-
-def test_trust_cli_macos_native_status_reports_setup_required_when_passive_backend_is_ready(
-    tmp_path: Path,
-    capsys,
-    monkeypatch: pytest.MonkeyPatch,
-    install_fake_system_keyring,
-) -> None:
-    _enable_macos_native_policy_integrity(monkeypatch, install_fake_system_keyring)
-    home_dir = tmp_path / "home"
-
-    rc = main(
-        [
-            "guard",
-            "trust",
-            "status",
-            "--backend",
-            "macos-native",
-            "--home",
-            str(home_dir),
-            "--json",
-        ]
-    )
-    payload = json.loads(capsys.readouterr().out)
-
-    assert rc == 0
-    assert payload["mode"] == "setup_required"
-    assert payload["backend_requested"] == "macos-native"
-    assert payload["backend_selected"] == "macos-native"
-    assert payload["setup_available"] is True
-    assert payload["passive_prompt_allowed"] is False
-
-
-def test_trust_cli_macos_native_test_stays_prompt_free_when_backend_is_unavailable(
-    tmp_path: Path,
-    capsys,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    home_dir = tmp_path / "home"
-    monkeypatch.setattr(trust_dispatch_module.sys, "platform", "darwin", raising=False)
-    monkeypatch.setattr(guard_store_module.sys, "platform", "darwin", raising=False)
-    monkeypatch.setattr(
-        SystemKeyringSecretStore,
-        "_supports_native_macos_security_reads",
-        classmethod(lambda cls: False),
-    )
-
-    rc = main(
-        [
-            "guard",
-            "trust",
-            "test",
-            "--backend",
-            "macos-native",
-            "--home",
-            str(home_dir),
-            "--no-ui",
-            "--json",
-        ]
-    )
-    payload = json.loads(capsys.readouterr().out)
-
-    assert rc == 0
-    assert payload["mode"] == "degraded_safe"
-    assert payload["ok"] is True
-    assert payload["trust_health"] == "degraded_safe"
-    assert payload["passive_prompt_allowed"] is False
-
-
 def test_trust_cli_doctor_degraded_safe_does_not_pass_runtime_check(tmp_path: Path, capsys) -> None:
     home_dir = tmp_path / "home"
     rc = main(
